campus/serializers.py
from django.contrib.auth.models import Group
from rest_framework import serializers
from .models import Post
from django.contrib.auth import get_user_model

MyUser = get_user_model()


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = MyUser
        fields = ['url', 'username', 'email', 'groups']


class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ['url', 'name']


class PostSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post 
        fields = ['content', 'author']