requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-countries==7.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-six==1.0.4
django-utils-six==2.0
djangorestframework==3.12.2
fontawesome-free==5.15.2
phonenumbers==8.12.19
Pillow==8.1.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
